+++ packages/node-core/src/indexer/worker/worker.monitor.service.ts
@@ -7,28 +7,30 @@ import {setMonitorService} from '../../process';
 import {MonitorServiceInterface} from '../monitor.service';
 
 type HostMonitorService = {
-  hostMonitorServiceWrite: (blockData: string) => void;
+  hostMonitorServiceWrite?: (blockData: string) => void;
 };
 
 export const hostMonitorKeys: (keyof HostMonitorService)[] = ['hostMonitorServiceWrite'];
 
-export function monitorHostFunctions(host: MonitorServiceInterface): HostMonitorService {
+export function monitorHostFunctions(host?: MonitorServiceInterface): HostMonitorService {
   return {
-    hostMonitorServiceWrite: host.write.bind(host),
+    hostMonitorServiceWrite: host?.write.bind(host),
   };
 }
 
 @Injectable()
 export class WorkerMonitorService implements MonitorServiceInterface {
-  constructor(private host: HostMonitorService) {
+  constructor(private host?: HostMonitorService) {
     if (isMainThread) {
       throw new Error('Expected to be worker thread');
     }
     setMonitorService(this);
   }
 
-  write(blockData: string): void {
-    return this.host?.hostMonitorServiceWrite ? this.host.hostMonitorServiceWrite(blockData) : undefined;
+  write(blockData: string | (() => string)): void {
+    return this.host?.hostMonitorServiceWrite
+      ? this.host.hostMonitorServiceWrite(typeof blockData === 'string' ? blockData : blockData())
+      : undefined;
   }
 
   createBlockFork(blockHeight: number): void {